Preserved onVisibleMonthsChange data to loadItemsForMonth
For some reason only the first month of the visible months was being passed on to the loadItemsForMonth property.  It's not currently clear to me why AgendaView would swallow that information when CalendarList is providing it.  This check in might break examples, but I want to confirm behavior before updating everything else.